requirements.txt
playwright>=1.40.0
pytest>=7.4.0
python-dotenv>=1.0.0
allure-pytest>=2.13.0
